refactor(spock): rename groovy-spock module to spock
- Rename module directory: groovy-spock → spock

- Update package: groovyspock → gvy.spock

- Update dependencies in gls and groovy-testing modules

- Update imports across dependent modules
