+++ settings.gradle.kts
@@ -33,7 +33,7 @@ include("diagnostics:sarif")
 include("jenkins")
 include("groovy-gdsl")
 include("build-tool")
-include("groovy-spock")
+include("spock")
 include("groovy-testing")
 include("groovy-test-utils")
 include("groovy-junit")
