+++ spock/build.gradle.kts
