+++ spock/src/test/kotlin/com/github/albertocavalcante/gvy/spock/SpockDetectorTest.kt
@@ -1,4 +1,4 @@
-package com.github.albertocavalcante.groovyspock
+package com.github.albertocavalcante.gvy.spock
 
 import org.codehaus.groovy.ast.ClassHelper
 import org.codehaus.groovy.ast.ClassNode
