+++ spock/src/test/kotlin/com/github/albertocavalcante/gvy/spock/SpockFeatureExtractorTest.kt
@@ -1,4 +1,4 @@
-package com.github.albertocavalcante.groovyspock
+package com.github.albertocavalcante.gvy.spock
 
 import com.github.albertocavalcante.groovyparser.GroovyParserFacade
 import com.github.albertocavalcante.nativeapi.ParseRequest
